+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Fix: `HeaderNav`: Fix user menu position on mobile.
+- Fix: `ContributionCard`: Fix image dimensions on desktop.
+- Feature: `ContributionCard`: Add `largeInput` prop.
 
 ## [3.25.2] - 2021-07-09